requirements.txt
pandas>=2.0.0
python-dateutil>=2.8.0
tqdm>=4.65.0
PyYAML>=6.0
actualpy
